switch Long to int

+++ src/main/java/com/yein/domain/Todo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class Todo {
-	private Long no;
+	private int no;
 	private String title;
 	private String content;
 	private Date regdate;
@@ -15,11 +15,11 @@ public class Todo {
 	public Todo() {
 	}
 
-	public Long getNo() {
+	public int getNo() {
 		return no;
 	}
 
-	public void setNo(Long no) {
+	public void setNo(int no) {
 		this.no = no;
 	}
